just some compiler warning cleanup.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2562 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_VideoDX9/Src/PixelShaderCache.cpp
@@ -85,7 +85,7 @@ void PixelShaderCache::SetShader()
 
 	bool HLSL = false;
 	const char *code = GeneratePixelShader(PixelShaderManager::GetTextureMask(), false, false, HLSL);
-	LPDIRECT3DPIXELSHADER9 shader = HLSL ? D3D::CompilePixelShader(code, strlen(code), false) : CompileCgShader(code);
+	LPDIRECT3DPIXELSHADER9 shader = HLSL ? D3D::CompilePixelShader(code, (int)strlen(code), false) : CompileCgShader(code);
 	if (shader)
 	{
 		//Make an entry in the table
